+++ src/main/java/org/openrewrite/java/spring/data/MigrateRepositoryRestConfigurerAdapter.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.openrewrite.java.spring.framework;
+package org.openrewrite.java.spring.data;
 
 import lombok.Getter;
 import org.jspecify.annotations.Nullable;
