+++ src/main/resources/META-INF/rewrite/recipes.csv
@@ -194,7 +194,7 @@ maven,org.openrewrite.recipe:rewrite-spring,org.openrewrite.java.spring.framewor
 maven,org.openrewrite.recipe:rewrite-spring,org.openrewrite.java.spring.framework.MigrateHandlerResultSetExceptionHandlerMethod,Migrate `org.springframework.web.reactive.HandlerResult.setExceptionHandler` method,"`org.springframework.web.reactive.HandlerResult.setExceptionHandler(Function<Throwable, Mono<HandlerResult>>)` was deprecated, in favor of `setExceptionHandler(DispatchExceptionHandler)`.",1,,,,Spring Framework,Spring,Java,,,,,Recipes for upgrading and patching [Spring](https://spring.io/) applications.,Basic building blocks for transforming Java code.,,
 maven,org.openrewrite.recipe:rewrite-spring,org.openrewrite.java.spring.framework.MigrateInstantiationAwareBeanPostProcessorAdapter,Convert `InstantiationAwareBeanPostProcessorAdapter` to `SmartInstantiationAwareBeanPostProcessor`,As of Spring-Framework 5.3 `InstantiationAwareBeanPostProcessorAdapter` is deprecated in favor of the existing default methods in `SmartInstantiationAwareBeanPostProcessor`.,1,,,,Spring Framework,Spring,Java,,,,,Recipes for upgrading and patching [Spring](https://spring.io/) applications.,Basic building blocks for transforming Java code.,,
 maven,org.openrewrite.recipe:rewrite-spring,org.openrewrite.java.spring.framework.MigrateMethodArgumentNotValidExceptionErrorMethod,Migrate `MethodArgumentNotValidException.errorsToStringList` and `resolveErrorMessages`,"`org.springframework.web.bind.MethodArgumentNotValidException.errorsToStringList` and `resolveErrorMessages` method was deprecated, in favor of `BindErrorUtils`.",1,,,,Spring Framework,Spring,Java,,,,,Recipes for upgrading and patching [Spring](https://spring.io/) applications.,Basic building blocks for transforming Java code.,,
-maven,org.openrewrite.recipe:rewrite-spring,org.openrewrite.java.spring.framework.MigrateRepositoryRestConfigurerAdapter,Replace `RepositoryRestConfigurerAdapter` with `RepositoryRestConfigurer`,Since 3.1 implement RepositoryRestConfigurer directly.,1,,,,Spring Framework,Spring,Java,,,,,Recipes for upgrading and patching [Spring](https://spring.io/) applications.,Basic building blocks for transforming Java code.,,
+maven,org.openrewrite.recipe:rewrite-spring,org.openrewrite.java.spring.data.MigrateRepositoryRestConfigurerAdapter,Replace `RepositoryRestConfigurerAdapter` with `RepositoryRestConfigurer`,Since 3.1 implement RepositoryRestConfigurer directly.,1,,,,Spring Framework,Spring,Java,,,,,Recipes for upgrading and patching [Spring](https://spring.io/) applications.,Basic building blocks for transforming Java code.,,
 maven,org.openrewrite.recipe:rewrite-spring,org.openrewrite.java.spring.framework.MigrateResourceHttpMessageWriterAddHeadersMethod,Migrate `ResourceHttpMessageWriter.addHeaders`,"`org.springframework.http.codec.ResourceHttpMessageWriter.addHeaders` was deprecated, in favor of `addDefaultHeaders` method.",1,,,,Spring Framework,Spring,Java,,,,,Recipes for upgrading and patching [Spring](https://spring.io/) applications.,Basic building blocks for transforming Java code.,,
 maven,org.openrewrite.recipe:rewrite-spring,org.openrewrite.java.spring.framework.MigrateResponseEntityExceptionHandlerHttpStatusToHttpStatusCode,Migrate `ResponseEntityExceptionHandler` from HttpStatus to HttpStatusCode,With Spring 6 `HttpStatus` was replaced by `HttpStatusCode` in most method signatures in the `ResponseEntityExceptionHandler`.,1,,,,Spring Framework,Spring,Java,,,,,Recipes for upgrading and patching [Spring](https://spring.io/) applications.,Basic building blocks for transforming Java code.,,
 maven,org.openrewrite.recipe:rewrite-spring,org.openrewrite.java.spring.framework.MigrateResponseStatusExceptionGetRawStatusCodeMethod,Migrate `ResponseStatusException#getRawStatusCode()` to `getStatusCode().value()`,Migrate Spring Framework 5.3's `ResponseStatusException` method `getRawStatusCode()` to Spring Framework 6's `getStatusCode().value()`.,1,,,,Spring Framework,Spring,Java,,,,,Recipes for upgrading and patching [Spring](https://spring.io/) applications.,Basic building blocks for transforming Java code.,,
