+++ src/main/resources/META-INF/rewrite/spring-boot-31.yml
@@ -53,6 +53,7 @@ recipeList:
       newVersion: 3.1.x
   - org.openrewrite.java.spring.security6.UpgradeSpringSecurity_6_1
   - org.openrewrite.java.spring.boot3.SpringBootProperties_3_1
+  - org.openrewrite.java.spring.data.MigrateRepositoryRestConfigurerAdapter
   - org.openrewrite.java.dependencies.UpgradeDependencyVersion:
       groupId: org.springdoc
       artifactId: "*"
@@ -60,4 +61,3 @@ recipeList:
   - org.openrewrite.hibernate.MigrateToHibernate62
   - org.openrewrite.java.testing.mockito.Mockito4to5Only
   - org.openrewrite.java.springdoc.UpgradeSpringDoc_2_2
-  - org.openrewrite.java.spring.framework.MigrateRepositoryRestConfigurerAdapter
