+++ src/test/java/org/openrewrite/java/spring/data/MigrateRepositoryRestConfigurerAdapterTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.openrewrite.java.spring.framework;
+package org.openrewrite.java.spring.data;
 
 import org.junit.jupiter.api.Test;
 import org.openrewrite.DocumentExample;
